+++ NEWS.md
@@ -4,6 +4,12 @@ NEW FUNCTIONS
 
 * `contr.deviation()` for sum-deviation contrast coding of factors.
 
+* `rowmean_n()`, to compute row means if row contains at least `n` non-missing
+  values.
+
+* `means_by_group()`, to compute mean values of variables, grouped by levels
+  of specified factors.
+
 CHANGES
 
 * `recode_into()` gains an `overwrite` argument to skip overwriting already